Die when we try to write a packfile that we cannot read.
Currently the packfile format simply does not allow for the data
it stores to contain a newline. Instead of allowing us to write
a broken packfile and then find out later it is unreadable we throw
an exception when we try to write.

This is part of a response to Issue #8.