Avoid un-necessary calls to `cwd()`, use `File::Spec->catfile(..)` instead.
While `File::Find` will `chdir()` to various directories while
traversing the list of files to install, we don't need to call `cwd()`
for each file; we already know "what directory we started in" along with
"what directory did `File::File` `chdir()` into", which we can combine
to get "the current directory".

While for most installs this is not of much significance, `Paws`
contains over 20,000 `.pm` files needing to be installed and this patch
allows us to eliminate 20,000 calls to `cwd()`, making the install
notably faster.